fix VS-2010 64bits warnings

+++ src/DGtal/io/viewers/Viewer3D.cpp
@@ -90,12 +90,12 @@ DGtal::Viewer3D::drawWithNames()
     }
   for(unsigned int i=0; i<myLineSetList.size(); i++)
     {
-      glCallList(myListToAff+myVoxelSetList.size()+i);
+      glCallList(GLuint(myListToAff+myVoxelSetList.size()+i));
     }
   
   for(unsigned int i=0; i<myPointSetList.size(); i++)
     {
-      glCallList(myListToAff+myVoxelSetList.size()+myLineSetList.size()+i);
+      glCallList(GLuint(myListToAff+myVoxelSetList.size()+myLineSetList.size()+i));
     }   
 }
 
@@ -130,7 +130,7 @@ DGtal::Viewer3D::draw()
       if(myPointSetList.at(i).size()!=0){
 	glPointSize((myPointSetList.at(i).at(0).size)/distCam);
       }
-      glCallList(myListToAff+myVoxelSetList.size()+myLineSetList.size()+i+1);
+      glCallList(GLuint(myListToAff+myVoxelSetList.size()+myLineSetList.size()+i+1));
     }   
  
   for(unsigned int i=0; i<myLineSetList.size(); i++)
@@ -138,10 +138,10 @@ DGtal::Viewer3D::draw()
       if(myLineSetList.at(i).size()!=0){
 	glLineWidth((myLineSetList.at(i).at(0).width));
       }
-      glCallList(myListToAff+myVoxelSetList.size()+1+i);
+      glCallList(GLuint(myListToAff+myVoxelSetList.size()+1+i));
     }
   
-  glCallList(myListToAff+myVoxelSetList.size());
+  glCallList(GLuint(myListToAff+myVoxelSetList.size()));
   for(unsigned int i=0; i<myVoxelSetList.size(); i++)
     {
       glCallList(myListToAff+i);
@@ -301,7 +301,7 @@ DGtal::Viewer3D::postSelection(const QPoint& point)
 void
 DGtal::Viewer3D::updateList(bool needToUpdateBoundingBox)
 { 
-  unsigned int nbList= myVoxelSetList.size()+ myLineSetList.size()+ myPointSetList.size();
+  unsigned int nbList= (unsigned int)(myVoxelSetList.size()+ myLineSetList.size()+ myPointSetList.size());
   glDeleteLists(myListToAff, myNbListe);
   myListToAff = glGenLists( nbList  );   
   myNbListe=0;
@@ -372,7 +372,7 @@ DGtal::Viewer3D::updateList(bool needToUpdateBoundingBox)
       glEnd();
       glEndList();
     }
-  glNewList(myListToAff+myVoxelSetList.size(), GL_COMPILE);
+  glNewList(GLuint(myListToAff+myVoxelSetList.size()), GL_COMPILE);
   myNbListe++;
   glPushName(myNbListe);  
   glEnable( GL_DEPTH_TEST );
@@ -402,7 +402,7 @@ DGtal::Viewer3D::updateList(bool needToUpdateBoundingBox)
   for (unsigned int i=0; i<myLineSetList.size(); i++)
     {  
       listeID++;
-      glNewList(myListToAff+myVoxelSetList.size()+i+1, GL_COMPILE);
+      glNewList(GLuint(myListToAff+myVoxelSetList.size()+i+1), GL_COMPILE);
       myNbListe++;
       glDisable(GL_LIGHTING);
       glPushName(myNbListe);  
@@ -425,7 +425,7 @@ DGtal::Viewer3D::updateList(bool needToUpdateBoundingBox)
 
   for (unsigned int i=0; i<myPointSetList.size(); i++)
     {  
-      glNewList(myListToAff+myLineSetList.size()+myVoxelSetList.size()+i+1, GL_COMPILE);
+      glNewList(GLuint(myListToAff+myLineSetList.size()+myVoxelSetList.size()+i+1), GL_COMPILE);
       myNbListe++;
       glDepthMask(GL_TRUE);
       glDisable(GL_TEXTURE_2D);
